+++ openghg/standardise/surface/_icos.py
@@ -17,7 +17,7 @@ def parse_icos(
     sampling_period: Optional[str] = None,
     measurement_type: Optional[str] = None,
     header_type: str = "large",
-    update_mismatch: bool = False,
+    update_mismatch: str = "never",
     site_filepath: optionalPathType = None,
     **kwargs: Dict,
 ) -> Dict:
@@ -33,10 +33,12 @@ def parse_icos(
         measurement_type: Measurement type e.g. insitu, flask
         header_type: ICOS data file with large (40 line) header or shorter single line header
             Options: large, small
-        update_mismatch: This determines whether mismatches between the internal data
-            attributes and the supplied / derived metadata can be updated or whether
-            this should raise an AttrMismatchError.
-            If True, currently updates metadata with attribute value.        
+        update_mismatch: This determines how mismatches between the internal data
+            "attributes" and the supplied / derived "metadata" are handled.
+            This includes the options:
+              - "never" - don't update mismatches and raise an AttrMismatchError
+              - "from_source" / "attributes" - update mismatches based on input data (e.g. data attributes)
+              - "from_definition" / "metadata" - update mismatches based on associated data (e.g. site_info.json)
         site_filepath: Alternative site info file (see openghg/supplementary_data repository for format).
             Otherwise will use the data stored within openghg_defs/data/site_info JSON file by default.
     Returns:
@@ -81,11 +83,13 @@ def parse_icos(
         )
 
     # Ensure the data is CF compliant
-    gas_data = assign_attributes(data=gas_data,
-                                 site=site,
-                                 sampling_period=sampling_period,
-                                 update_mismatch=update_mismatch,
-                                 site_filepath=site_filepath)
+    gas_data = assign_attributes(
+        data=gas_data,
+        site=site,
+        sampling_period=sampling_period,
+        update_mismatch=update_mismatch,
+        site_filepath=site_filepath,
+    )
 
     return gas_data