+++ openghg/store/_obssurface.py
@@ -114,14 +114,13 @@ class ObsSurface(BaseStore):
         source_format: str,
         network: str,
         site: str,
-        store: Optional[str] = None,
         inlet: Optional[str] = None,
         height: Optional[str] = None,
         instrument: Optional[str] = None,
         sampling_period: Optional[Union[Timedelta, str]] = None,
         calibration_scale: Optional[str] = None,
         measurement_type: str = "insitu",
-        update_mismatch: bool = False,
+        update_mismatch: str = "never",
         overwrite: bool = False,
         verify_site_code: bool = True,
         site_filepath: optionalPathType = None,
@@ -143,14 +142,16 @@ class ObsSurface(BaseStore):
             instrument: Instrument name
             sampling_period: Sampling period in pandas style (e.g. 2H for 2 hour period, 2m for 2 minute period).
             measurement_type: Type of measurement e.g. insitu, flask
-            update_mismatch: This determines whether mismatches between the internal data
-                attributes and the supplied / derived metadata can be updated or whether
-                this should raise an AttrMismatchError.
-                If True, currently updates metadata with attribute value.
+            update_mismatch: This determines how mismatches between the internal data
+                  "attributes" and the supplied / derived "metadata" are handled.
+                  This includes the options:
+                      - "never" - don't update mismatches and raise an AttrMismatchError
+                      - "from_source" / "attributes" - update mismatches based on input data (e.g. data attributes)
+                      - "from_definition" / "metadata" - update mismatches based on associated data (e.g. site_info.json)
             overwrite: Overwrite previously uploaded data
-            verify_site_code: Verify the site code
-            site_filepath: Alternative site info file (see openghg/supplementary_data repository for format).
-                Otherwise will use the data stored within openghg_defs/data/site_info JSON file by default.
+                  verify_site_code: Verify the site code
+                  site_filepath: Alternative site info file (see openghg/supplementary_data repository for format).
+                      Otherwise will use the data stored within openghg_defs/data/site_info JSON file by default.
         Returns:
             dict: Dictionary of Datasource UUIDs