+++ tests/dataobjects/test_datamanager.py
@@ -203,6 +203,9 @@ def test_delete_metadata_keys():
         "station_latitude": 52.51775,
         "station_long_name": "tacolneston tower, uk",
         "station_height_masl": 50.0,
+        "start_date": "2012-07-31 14:50:30+00:00",
+        "end_date": "2019-06-26 15:54:29+00:00",
+        "latest_version": "v1",
         "uuid": "test-uuid-100",
     }